cheri: Fix compile and warnings in drivers/net/ethernet/
- Standard uintptr_t vs. unsigned long issues.
- Some nasty hacks are required in drivers/net/ti/.
- neterion and 3com drivers must fabricate pointers from
  opaque data in hardware structures.

+++ drivers/net/ethernet/8390/8390.h
@@ -102,7 +102,7 @@ struct ei_device {
 	unsigned char saved_irq;	/* Original dev->irq value. */
 	u32 *reg_offset;		/* Register mapping table */
 	spinlock_t page_lock;		/* Page register locks */
-	unsigned long priv;		/* Private field to store bus IDs etc. */
+	uintptr_t priv;			/* Private field to store bus IDs etc. */
 	u32 msg_enable;			/* debug message level */
 #ifdef AX88796_PLATFORM
 	unsigned char rxcr_base;	/* default value for RXCR */
