+++ drivers/net/ethernet/intel/ice/ice_flow.h
@@ -462,7 +462,7 @@ struct ice_flow_entry {
 	u16 vsi_handle;
 };
 
-#define ICE_FLOW_ENTRY_HNDL(e)	((u64)(uintptr_t)e)
+#define ICE_FLOW_ENTRY_HNDL(e)	((uintptr_t)e)
 #define ICE_FLOW_ENTRY_PTR(h)	((struct ice_flow_entry *)(uintptr_t)(h))
 
 struct ice_flow_prof {
@@ -508,8 +508,8 @@ ice_flow_set_parser_prof(struct ice_hw *hw, u16 dest_vsi, u16 fdir_vsi,
 int
 ice_flow_add_entry(struct ice_hw *hw, enum ice_block blk, u64 prof_id,
 		   u64 entry_id, u16 vsi, enum ice_flow_priority prio,
-		   void *data, u64 *entry_h);
-int ice_flow_rem_entry(struct ice_hw *hw, enum ice_block blk, u64 entry_h);
+		   void *data, uintptr_t *entry_h);
+int ice_flow_rem_entry(struct ice_hw *hw, enum ice_block blk, uintptr_t entry_h);
 void
 ice_flow_set_fld(struct ice_flow_seg_info *seg, enum ice_flow_field fld,
 		 u16 val_loc, u16 mask_loc, u16 last_loc, bool range);
